Properly handle conflicting changes
Simultaneous edits are now directly merged (so-called "Fast merge") if
only different files have been changed. If the same file was touched in
both ancestry lines, a "conflict branch" is created for later merging. The
user is redirected to a site that allows him to do this merge (not yet working).